Readme
This To-Do List web page is built using HTML, CSS, and JavaScript. It’s crafted to be straightforward and intuitive, enabling users to effortlessly add, mark as complete, and delete tasks. The integration of HTML, CSS, and JavaScript delivers a fully operational and aesthetically pleasing To-Do List application, providing users with an effective way to organize and manage their tasks.